requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pandas==2.2.2
numpy==1.26.4
supabase==2.15.0
google-generativeai==0.8.5
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.7.1
google-cloud-texttospeech>=2.16.0
